trie

+++ codes/String/RWayTrie.cpp
@@ -136,10 +136,15 @@ class Trie
 				// POINTING to LAST character of the
 				// word we makes it value -1 as this word
 				// is no longer part of Trie obj
-				x->next[c]->value = -1;
-
-				// x->next[c] node is now checked whether it can be deleted
-				(*flag) = deleteGivenNode(&(x->next[c]));
+				if (root->next[c]!=NULL)
+				{
+					if (root->next[c]->value != -1)
+					{
+						x->next[c]->value = -1;
+						// x->next[c] node is now checked whether it can be deleted
+						(*flag) = deleteGivenNode(&(x->next[c]));	
+					}
+				}
 				return x;
 			}
 			int c = charAtd(key,d);
